Updated User Service Error Handling
- updated user service search error handling - now trap for missing params and return appropriate error response/message
- moved models to separate file
- cleaned up output
