+++ cla-backend-go/users/models.go
@@ -1,0 +1,19 @@
+// Copyright The Linux Foundation and each contributor to CommunityBridge.
+// SPDX-License-Identifier: MIT
+
+package users
+
+// DBUser data model
+type DBUser struct {
+	UserID             string   `json:"user_id"`
+	LFEmail            string   `json:"lf_email"`
+	LFUsername         string   `json:"lf_username"`
+	DateCreated        string   `json:"date_created"`
+	DateModified       string   `json:"date_modified"`
+	UserName           string   `json:"user_name"`
+	Version            string   `json:"version"`
+	UserEmails         []string `json:"user_emails"`
+	UserGithubID       string   `json:"user_github_id"`
+	UserCompanyID      string   `json:"user_company_id"`
+	UserGithubUsername string   `json:"user_github_username"`
+}
